+++ src/views/AboutView.vue
@@ -1,48 +1,4 @@
-<script setup lang="ts">
-import BoardDisplay from '@/components/BoardDisplayV2.vue';
-import {
-  ConfigOrderType,
-  getValidSquaresForOrder,
-} from '@/game/zugzwang/validators';
-import { ORDER_PRIORITIES } from '@/game/orders';
-import { coordinatesToArray } from '@/game/common';
-
-const orderNames: ConfigOrderType[] = [
-  'move-straight',
-  'move-diagonal',
-  'push-straight',
-  'push-diagonal',
-];
-
-const position = { x: 1, y: 1 };
-
-const pieces = [{ id: 0, owner: 0, position, isDefending: false }];
-
-const board = { x: 3, y: 3 };
-const fullBoard = { x: 4, y: 4 };
-
-let ordersDict = {};
-
-orderNames.forEach((name) => {
-  ordersDict[name] = getValidSquaresForOrder({
-    playerID: 0,
-    board,
-    orderType: name,
-  }).map((coord) => ({
-    type: name,
-    sourcePieceId: 0,
-    toTarget: coord,
-    priority: 1,
-    owner: 0,
-  }));
-});
-
-const placeActionSquares = getValidSquaresForOrder({
-  playerID: 0,
-  board: fullBoard,
-  orderType: 'place',
-}).map((coord) => coordinatesToArray(coord, fullBoard));
-</script>
+<script setup lang="ts"></script>
 
 <template>
   <main class="about">
@@ -64,59 +20,10 @@ const placeActionSquares = getValidSquaresForOrder({
         good in its own right before moving on. I hope you like it!
       </p>
     </section>
-    <br />
-    <h2>Order configuration:</h2>
-    <section class="grid">
-      <template v-for="order in orderNames" :key="order">
-        <p>{{ order }}</p>
-        <BoardDisplay
-          :pieces="pieces"
-          :board="board"
-          :orders="ordersDict[order]"
-        />
-        <div>
-          <p>priority: {{ ORDER_PRIORITIES[order] }}</p>
-          <p v-if="order === 'move-straight'">
-            "move" actions are designed to always advance toward the opponent's
-            side. If the target square of a "move" action is occupied when it's
-            supposed to occur, nothing happens.
-          </p>
-          <p v-if="order === 'push-straight'">
-            "push" actions leave the acting piece in place. This action pushes a
-            piece on the target square away, along with any others in the path.
-            Think of a sliding puzzle. Any pieces pushed off the board are
-            destroyed.
-          </p>
-        </div>
-      </template>
-      <p>place</p>
-      <BoardDisplay
-        :pieces="[]"
-        :orders="[]"
-        :board="{ x: 4, y: 4 }"
-        :highlighted-cells="placeActionSquares"
-      />
-      <div>
-        <p>priority: {{ ORDER_PRIORITIES.place }}</p>
-        <p>
-          For each of your pieces that are missing, you may place a new piece on
-          your home row during your turn. If there is a piece in that square,
-          <strong>both will be destroyed!</strong>
-        </p>
-      </div>
-    </section>
   </main>
 </template>
 
 <style>
-strong {
-  font-weight: bold;
-}
-.grid {
-  display: grid;
-  grid-template-columns: 1fr 1fr 1fr;
-  row-gap: 1rem;
-}
 .text {
   max-width: 800px;
   margin: 0 auto;
